Move contexts to dedicated folder
