+++ src/auth/LogoutButton.jsx
@@ -1,6 +1,6 @@
 // src/auth/LogoutButton.jsx
 import React from "react";
-import { useAuth } from "./AuthContext";
+import { useAuth } from "../contexts/AuthContext";
 
 export default function LogoutButton() {
   const { logout } = useAuth();
