+++ src/auth/OAuthCallback.jsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "./AuthContext";
+import { useAuth } from "../contexts/AuthContext";
 
 export default function OAuthCallback() {
   const navigate = useNavigate();
