+++ src/components/buttons/ShareButton.jsx
@@ -1,8 +1,8 @@
 // src/components/buttons/ShareButton.jsx
 import React from "react";
 import SvgShare from "../svgs/SvgShare";
-import { useAuth } from "../../auth/AuthContext";
-import { useSidebar } from "../../ui/SidebarContext";
+import { useAuth } from "../../contexts/AuthContext";
+import { useSidebar } from "../../contexts/SidebarContext";
 
 export default function ShareButton({ title, url }) {
   const { user } = useAuth();
